requirements.txt
Pillow==6.0.0
pyTelegramBotAPI==3.7.9
requests==2.19.1